requirements.txt
Flask==2.1.2
PyJWT==2.4.0
oauth2client==4.1.3
google-auth-oauthlib==0.5.2
google-api-python-client==2.52.0
pymongo==4.1.1
python-dotenv==0.20.0
Flask-Dance==6.0.0
gunicorn==20.0.4
dnspython==2.2.1
boto3==1.24.37
Flask-Cors==3.0.10